requirements.txt
streamlit>=1.32.0
pandas>=2.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=5.1.0
python-dotenv>=1.0.0
plotly>=5.18.0
tenacity>=8.2.0
fake-useragent>=1.4.0
pytest>=7.4.0
